requirements.txt
Flask
Flask_SQLAlchemy
requests
psycopg2-binary
geoalchemy2
shapely